requirements.txt
boto3==1.18.18
ffmpeg-python==0.2.0
Flask==2.0.1
numpy==1.21.1
pymongo==3.12.0
scipy==1.7.1
speechbrain==0.5.9
Werkzeug==2.0.1